backend/requirements.txt

# Core (cài trước)
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
starlette==0.48.0

# ===================================
# SETTINGS & CONFIGURATION
# ===================================
pydantic==2.12.0
pydantic-settings==2.11.0
python-dotenv==1.1.1
email-validator==2.3.0

# ===================================
# DATABASE
# ===================================
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.0
greenlet==3.2.4

# ===================================
# SECURITY & AUTHENTICATION
# ===================================
python-jose[cryptography]==3.5.0
argon2-cffi==25.1.0
passlib[argon2]==1.7.4
bcrypt==4.0.1
cryptography==46.0.2

# ===================================
# AI & LLM 
# ===================================
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.71.0

# Langchain 
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11

# ===================================
# FILE PROCESSING & EXTRACTION
# ===================================
PyPDF2==3.0.1           # For PDF extraction (light ~5MB)
python-docx==1.2.0      # For DOCX extraction (light ~3MB)
pillow==11.3.0          # Image support (if needed)

# ===================================
# VECTOR DATABASE
# ===================================
pinecone==7.3.0

# ===================================
# HTTP & NETWORKING
# ===================================
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
aiohttp==3.13.0

# ===================================
# UTILITIES
# ===================================
python-dateutil==2.9.0.post0
PyYAML==6.0.3
typing-extensions==4.15.0
packaging==24.2
certifi==2025.10.5
charset-normalizer==3.4.3

# ===================================
# Export features
# ===================================
reportlab==4.0.7
python-pptx==0.6.21
fpdf==1.7.2
openpyxl==3.1.2